+++ homeassistant/components/sabnzbd/config_flow.py
@@ -16,7 +16,7 @@ from homeassistant.helpers.selector import (
 )
 
 from .const import DOMAIN
-from .sab import get_client
+from .helpers import get_client
 
 _LOGGER = logging.getLogger(__name__)
